kbuild: Increase kallsyms max symbol length
[AK: This seems like a ticking time bomb even without LTO,
so should be merged now. It causes very weird problems.
Thanks to Joe for tracking them down.]

With the added postfixes that LTO adds for local
symbols, the longest name in the kernel overflows
the namebuf[KSYM_NAME_LEN] array by two bytes.  That name is:
__pci_fixup_resumePCI_VENDOR_ID_SERVERWORKSPCI_DEVICE_ID_SERVERWORKS_HT1000SBquirk_disable_broadcom_boot_interrupt.1488004.672802

Double the max symbol name length.

Change-Id: I19eca9a97333b90148a58d3e3c10344b2c0922e7
Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: John Vincent <git@tensevntysevn.cf>
Signed-off-by: John Vincent <git@tenseventyseven.cf>
Signed-off-by: fadlyas07 <mhmmdfdlyas@gmail.com>
Signed-off-by: HeroBuxx <me@herobuxx.me>